tools update

+++ scripts/remove/tools.sh
@@ -2,9 +2,23 @@
 
 #rm -rf /srv/tools
 rm /srv/tools/*.*
-rm -rf /srv/tools/vendor
 
-rm -f /etc/nginx/apps/tools.conf
-rm -f /etc/sudoers.d/tools
-rm /etc/cron.d/set_interface_tools
-rm /install/.tools.lock
+if [[ ! -f /srv/tools/vendor ]]; then
+  rm -rf /srv/tools/vendor
+fi
+
+if [[ ! -f /etc/nginx/apps/tools.conf ]]; then
+  rm -f /etc/nginx/apps/tools.conf
+fi
+
+if [[ ! -f /etc/sudoers.d/tools ]]; then
+  rm -f /etc/sudoers.d/tools
+fi
+
+if [[ ! -f /etc/cron.d/set_interface_tools ]]; then
+  rm /etc/cron.d/set_interface_tools
+fi
+
+if [[ ! -f /install/.tools.lock ]]; then
+  rm /install/.tools.lock
+fi
